requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
SpeechRecognition
pydub
eng_to_ipa
requests
